+++ support/apps/apps.go
@@ -19,30 +19,41 @@ import (
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
+// ExtractVersion extracts version info from given command
+func ExtractVersion(cmd string, line, field int) support.App {
+	if len(cmd) == 0 {
+		return support.App{}
+	}
+
+	cmdSlice := strings.Fields(cmd)
+
+	return support.App{cmdSlice[0], extractField(execVersionCmd(cmdSlice...), line, field)}
+}
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
 // Golang extracts version info from go command output
 func Golang() support.App {
-	ver := extractField(execVersionCmd("go", "version"), 0, 2)
-	ver = strutil.Exclude(ver, "go")
+	info := ExtractVersion("go version", 0, 2)
 
-	return support.App{"golang", ver}
+	info.Version = strutil.Exclude(info.Version, "go")
+
+	return info
 }
 
 // GCC extracts version info from gcc command output
 func GCC() support.App {
-	ver := extractField(execVersionCmd("gcc", "--version"), 0, 2)
-	return support.App{"gcc", ver}
+	return ExtractVersion("gcc --version", 0, 2)
 }
 
 // Clang extracts version info from clang command output
 func Clang() support.App {
-	ver := extractField(execVersionCmd("clang", "--version"), 0, 2)
-	return support.App{"clang", ver}
+	return ExtractVersion("clang --version", 0, 2)
 }
 
 // Python3 extracts version info from python 3.x command output
 func Python3() support.App {
-	ver := extractField(execVersionCmd("python3", "--version"), 0, 1)
-	return support.App{"python3", ver}
+	return ExtractVersion("python3 --version", 0, 1)
 }
 
 // Java extracts version info from java command output
@@ -60,85 +71,78 @@ func Java() support.App {
 
 // Groovy extracts version info from groovy command output
 func Groovy() support.App {
-	ver := extractField(execVersionCmd("groovy", "--version"), 0, 2)
-	return support.App{"groovy", ver}
+	return ExtractVersion("groovy --version", 0, 2)
 }
 
 // Ruby extracts version info from ruby command output
 func Ruby() support.App {
-	ver := extractField(execVersionCmd("ruby", "--version"), 0, 1)
-	return support.App{"ruby", ver}
+	return ExtractVersion("ruby --version", 0, 1)
 }
 
 // NodeJS extracts version info from node command output
 func NodeJS() support.App {
-	ver := extractField(execVersionCmd("node", "--version"), 0, 0)
-	ver = strings.TrimLeft(ver, "v")
+	info := ExtractVersion("node --version", 0, 2)
+	info.Version = strings.TrimLeft(info.Version, "v")
 
-	return support.App{"nodejs", ver}
+	return info
 }
 
 // Rust extracts version info from rust command output
 func Rust() support.App {
-	ver := extractField(execVersionCmd("rustc", "--version"), 0, 1)
-	return support.App{"rust", ver}
+	return ExtractVersion("rustc --version", 0, 1)
 }
 
 // PHP extracts version info from php command output
 func PHP() support.App {
-	ver := extractField(execVersionCmd("php", "--version"), 0, 1)
-	return support.App{"php", ver}
+	return ExtractVersion("php --version", 0, 1)
 }
 
 // Bash extracts version info from bash command output
 func Bash() support.App {
-	ver := extractField(execVersionCmd("bash", "--version"), 0, 3)
-	return support.App{"bash", ver}
+	return ExtractVersion("bash --version", 0, 3)
 }
 
 // Git extracts version info from git command output
 func Git() support.App {
-	ver := extractField(execVersionCmd("git", "--version"), 0, 2)
-	return support.App{"git", ver}
+	return ExtractVersion("git --version", 0, 2)
 }
 
 // Mercurial extracts version info from hg command output
 func Mercurial() support.App {
-	ver := extractField(execVersionCmd("hg", "--version"), 0, 4)
-	ver = strings.Trim(ver, "()")
+	info := ExtractVersion("hg --version", 0, 4)
+	info.Version = strings.Trim(info.Version, "()")
 
-	return support.App{"mercurial", ver}
+	return info
 }
 
 // SVN extracts version info from svn command output
 func SVN() support.App {
-	ver := extractField(execVersionCmd("svn", "--version"), 0, 2)
-	return support.App{"svn", ver}
+	return ExtractVersion("svn --version", 0, 2)
 }
 
 // Docker extracts version info from Docker command output
 func Docker() support.App {
-	ver := extractField(execVersionCmd("docker", "--version"), 0, 2)
-	return support.App{"docker", strings.TrimRight(ver, ",")}
+	info := ExtractVersion("docker --version", 0, 2)
+	info.Version = strings.TrimRight(info.Version, ",")
+
+	return info
 }
 
 // Podman extracts version info from Podman command output
 func Podman() support.App {
-	ver := extractField(execVersionCmd("podman", "--version"), 0, 2)
-	return support.App{"podman", ver}
+	return ExtractVersion("podman --version", 0, 2)
 }
 
 // LXC extracts version info from LXC command output
 func LXC() support.App {
-	ver := extractField(execVersionCmd("lxc", "--version"), 0, 0)
-	return support.App{"lxc", ver}
+	return ExtractVersion("lxc --version", 0, 0)
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 // execVersionCmd execs command and returns output as a string
-func execVersionCmd(name string, args ...string) string {
-	c := exec.Command(name, args...)
+func execVersionCmd(cmd ...string) string {
+	c := exec.Command(cmd[0], cmd[1:]...)
 	output, err := c.CombinedOutput()
 
 	if err != nil {