+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### [13.26.2](https://kaos.sh/ek/13.26.2)
+
+- **`[support]`** Improved handling empty slices with data (_deps, apps, services…_)
+- **`[support/apps]`** Added method `ExtractVersion` for extracting version info from command output
+
 ### [13.26.1](https://kaos.sh/ek/13.26.1)
 
 - **`[usage/update]`** Use token from `GH_TOKEN` environment variable for requests to Github API